Refactoring egg.js into its own component and simplifying fv-egg-clutch
