+++ src/code/templates/egg-sort-game.js
@@ -7,7 +7,8 @@ import initialElementId from '../hoc/initial-element-id';
 import targetElementId from '../hoc/target-element-id';
 import updateOnResizeScroll from '../hoc/update-on-resize-scroll';
 import BasketSetView from '../components/basket-set';
-import EggClutchView, { EGG_IMAGE_WIDTH } from '../components/egg-clutch';
+import EggClutchView from '../components/egg-clutch';
+import { EGG_IMAGE_WIDTH } from '../components/egg';
 import EggHatchView from '../components/egg-hatch';
 import GenomeView from '../components/genome';
 import { generateTrialDrakes } from '../utilities/trial-generator';
