+++ src/code/templates/fv-egg-sort-game.js
@@ -16,7 +16,8 @@ import initialElementId from '../hoc/initial-element-id';
 import targetElementId from '../hoc/target-element-id';
 import updateOnResizeScroll from '../hoc/update-on-resize-scroll';
 import BasketSetView from '../components/basket-set';
-import FVEggClutchView, { EGG_IMAGE_WIDTH } from '../fv-components/fv-egg-clutch';
+import FVEggClutchView from '../fv-components/fv-egg-clutch';
+import { EGG_IMAGE_WIDTH } from '../components/egg';
 import EggHatchView from '../components/egg-hatch';
 import GenomeView from '../components/genome';
 import FVChromosomeImageView from '../fv-components/fv-chromosome-image';
